v3.1.2 - Bugfixes and Refactoring. No new features.
Refactoring:
New introduction in Main that describes some to-be-implemented features
Reordered a bunch of code in Main.
Removed most magic numbers in Main.
Renamed a few fields in Main for readability.
Clarified which fields in the Main class should be declared outside of a function.
Corrected a spelling mistake in Drag.
Added three translation methods in CardImage to make operations on cards more intuitive.
Removed setDraggable from CardImage, it should only be set in the constructor.
Clarified the CardImage construction values for a blank card.
Modified Drag to use the new translation methods of CardImage where possible.
Added a few clarifying comments in various locations.
Gave up on trying to keep lines under 80 characters, for the sake of readability during production.

Bugfixes:
Fixed an issue where dragging a card onto an ace would throw an exception and fail to reset the dragged card if the ace was in the very first grid position.